+++ src/spaceone/identity/manager/service_account_manager.py
@@ -3,7 +3,7 @@ from typing import Tuple, List
 
 from spaceone.core.manager import BaseManager
 from spaceone.core.connector.space_connector import SpaceConnector
-from spaceone.identity.model.service_account_db import ServiceAccount
+from spaceone.identity.model.service_account.database import ServiceAccount
 
 _LOGGER = logging.getLogger(__name__)
